refactor: move product card to common
