+++ src/modules/common/components/ProductCard/index.ts
