+++ src/modules/overview/components/Movers/Movers.tsx
@@ -2,15 +2,15 @@ import { Box, ButtonBase, Grid, Typography, useTheme } from '@material-ui/core';
 import classNames from 'classnames';
 import { AngleLeftIcon } from 'modules/common/components/Icons/AngleLeftIcon';
 import { AngleRightIcon } from 'modules/common/components/Icons/AngleRightIcon';
+import {
+  IProductCardProps,
+  ProductCard,
+} from 'modules/common/components/ProductCard';
 import { QueryError } from 'modules/common/components/QueryError/QueryError';
 import { QueryLoadingCentered } from 'modules/common/components/QueryLoading/QueryLoading';
 import { SwiperPreloader } from 'modules/common/components/SwiperPreloader';
 import { getRandomId } from 'modules/common/utils/getRandomId';
 import { t } from 'modules/i18n/utils/intl';
-import {
-  IProductCardProps,
-  ProductCard,
-} from 'modules/overview/components/ProductCard';
 import { ISectionProps, Section } from 'modules/uiKit/Section';
 import React, { useEffect, useState } from 'react';
 import { uid } from 'react-uid';
